Rename files
Change file names from camelCase to kebab-case requesetd by mentor
