+++ src/js/modules/library-handlers.js
@@ -2,7 +2,7 @@ import { Notify } from 'notiflix';
 import { load, save, clear } from '../utils/storage';
 import { getLibraryMovies } from './get-library';
 import { generatePagination } from './pagination';
-import { renderMovies } from './renderItems';
+import { renderMovies } from './render-items';
 import { afterPaginationMove } from '../main';
 
 const configNotiflix = {
